+++ xds/src/test/java/io/grpc/xds/internal/security/certprovider/CertProviderServerSslContextProviderTest.java
@@ -14,113 +14,113 @@
  * limitations under the License.
  */
 
-package io.grpc.xds.internal.certprovider;
+package io.grpc.xds.internal.security.certprovider;
 
-import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.truth.Truth.assertThat;
-import static io.grpc.xds.internal.certprovider.CommonCertProviderTestUtils.getCertFromResourceName;
 import static io.grpc.xds.internal.security.CommonTlsContextTestsUtil.CA_PEM_FILE;
-import static io.grpc.xds.internal.security.CommonTlsContextTestsUtil.CLIENT_KEY_FILE;
 import static io.grpc.xds.internal.security.CommonTlsContextTestsUtil.CLIENT_PEM_FILE;
+import static io.grpc.xds.internal.security.CommonTlsContextTestsUtil.SERVER_0_KEY_FILE;
 import static io.grpc.xds.internal.security.CommonTlsContextTestsUtil.SERVER_0_PEM_FILE;
 import static io.grpc.xds.internal.security.CommonTlsContextTestsUtil.SERVER_1_KEY_FILE;
 import static io.grpc.xds.internal.security.CommonTlsContextTestsUtil.SERVER_1_PEM_FILE;
 import static io.grpc.xds.internal.security.CommonTlsContextTestsUtil.doChecksOnSslContext;
+import static io.grpc.xds.internal.security.certprovider.CommonCertProviderTestUtils.getCertFromResourceName;
 import static org.junit.Assert.fail;
 
-import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableList;
 import com.google.common.util.concurrent.MoreExecutors;
 import io.envoyproxy.envoy.config.core.v3.DataSource;
 import io.envoyproxy.envoy.extensions.transport_sockets.tls.v3.CertificateValidationContext;
+import io.envoyproxy.envoy.type.matcher.v3.StringMatcher;
 import io.grpc.xds.Bootstrapper;
 import io.grpc.xds.CommonBootstrapperTestUtils;
 import io.grpc.xds.EnvoyServerProtoData;
 import io.grpc.xds.internal.security.CommonTlsContextTestsUtil;
 import io.grpc.xds.internal.security.CommonTlsContextTestsUtil.TestCallback;
-import java.util.Queue;
-import java.util.concurrent.ConcurrentLinkedQueue;
-import java.util.concurrent.Executor;
-import java.util.logging.Level;
-import java.util.logging.Logger;
+import io.grpc.xds.internal.security.certprovider.CertProviderClientSslContextProviderTest.QueuedExecutor;
+import java.util.Arrays;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-/** Unit tests for {@link CertProviderClientSslContextProvider}. */
+/** Unit tests for {@link CertProviderServerSslContextProvider}. */
 @RunWith(JUnit4.class)
-public class CertProviderClientSslContextProviderTest {
-  private static final Logger logger =
-      Logger.getLogger(CertProviderClientSslContextProviderTest.class.getName());
+public class CertProviderServerSslContextProviderTest {
 
   CertificateProviderRegistry certificateProviderRegistry;
   CertificateProviderStore certificateProviderStore;
-  private CertProviderClientSslContextProvider.Factory certProviderClientSslContextProviderFactory;
+  private CertProviderServerSslContextProviderFactory certProviderServerSslContextProviderFactory;
 
   @Before
   public void setUp() throws Exception {
     certificateProviderRegistry = new CertificateProviderRegistry();
     certificateProviderStore = new CertificateProviderStore(certificateProviderRegistry);
-    certProviderClientSslContextProviderFactory =
-        new CertProviderClientSslContextProvider.Factory(certificateProviderStore);
+    certProviderServerSslContextProviderFactory =
+        new CertProviderServerSslContextProviderFactory(certificateProviderStore);
   }
 
-  /** Helper method to build CertProviderClientSslContextProvider. */
-  private CertProviderClientSslContextProvider getSslContextProvider(
+  /** Helper method to build CertProviderServerSslContextProvider. */
+  private CertProviderServerSslContextProvider getSslContextProvider(
       String certInstanceName,
       String rootInstanceName,
       Bootstrapper.BootstrapInfo bootstrapInfo,
       Iterable<String> alpnProtocols,
-      CertificateValidationContext staticCertValidationContext) {
-    EnvoyServerProtoData.UpstreamTlsContext upstreamTlsContext =
-        CommonTlsContextTestsUtil.buildUpstreamTlsContextForCertProviderInstance(
+      CertificateValidationContext staticCertValidationContext,
+      boolean requireClientCert) {
+    EnvoyServerProtoData.DownstreamTlsContext downstreamTlsContext =
+        CommonTlsContextTestsUtil.buildDownstreamTlsContextForCertProviderInstance(
             certInstanceName,
             "cert-default",
             rootInstanceName,
             "root-default",
             alpnProtocols,
-            staticCertValidationContext);
-    return certProviderClientSslContextProviderFactory.getProvider(
-        upstreamTlsContext,
+            staticCertValidationContext,
+            requireClientCert);
+    return certProviderServerSslContextProviderFactory.getProvider(
+        downstreamTlsContext,
         bootstrapInfo.node().toEnvoyProtoNode(),
         bootstrapInfo.certProviders());
   }
 
-  /** Helper method to build CertProviderClientSslContextProvider. */
-  private CertProviderClientSslContextProvider getNewSslContextProvider(
+  /** Helper method to build CertProviderServerSslContextProvider. */
+  private CertProviderServerSslContextProvider getNewSslContextProvider(
           String certInstanceName,
           String rootInstanceName,
           Bootstrapper.BootstrapInfo bootstrapInfo,
           Iterable<String> alpnProtocols,
-          CertificateValidationContext staticCertValidationContext) {
-    EnvoyServerProtoData.UpstreamTlsContext upstreamTlsContext =
-            CommonTlsContextTestsUtil.buildNewUpstreamTlsContextForCertProviderInstance(
+          CertificateValidationContext staticCertValidationContext,
+          boolean requireClientCert) {
+    EnvoyServerProtoData.DownstreamTlsContext downstreamTlsContext =
+            CommonTlsContextTestsUtil.buildNewDownstreamTlsContextForCertProviderInstance(
                     certInstanceName,
                     "cert-default",
                     rootInstanceName,
                     "root-default",
                     alpnProtocols,
-                    staticCertValidationContext);
-    return certProviderClientSslContextProviderFactory.getProvider(
-            upstreamTlsContext,
+                    staticCertValidationContext,
+                    requireClientCert);
+    return certProviderServerSslContextProviderFactory.getProvider(
+            downstreamTlsContext,
             bootstrapInfo.node().toEnvoyProtoNode(),
             bootstrapInfo.certProviders());
   }
 
+
   @Test
-  public void testProviderForClient_mtls() throws Exception {
+  public void testProviderForServer_mtls() throws Exception {
     final CertificateProvider.DistributorWatcher[] watcherCaptor =
         new CertificateProvider.DistributorWatcher[1];
     TestCertificateProvider.createAndRegisterProviderProvider(
         certificateProviderRegistry, watcherCaptor, "testca", 0);
-    CertProviderClientSslContextProvider provider =
+    CertProviderServerSslContextProvider provider =
         getSslContextProvider(
             "gcp_id",
             "gcp_id",
             CommonBootstrapperTestUtils.getTestBootstrapInfo(),
             /* alpnProtocols= */ null,
-            /* staticCertValidationContext= */ null);
+            /* staticCertValidationContext= */ null,
+            /* requireClientCert= */ true);
 
     assertThat(provider.savedKey).isNull();
     assertThat(provider.savedCertChain).isNull();
@@ -129,8 +129,8 @@ public class CertProviderClientSslContextProviderTest {
 
     // now generate cert update
     watcherCaptor[0].updateCertificate(
-        CommonCertProviderTestUtils.getPrivateKey(CLIENT_KEY_FILE),
-        ImmutableList.of(getCertFromResourceName(CLIENT_PEM_FILE)));
+        CommonCertProviderTestUtils.getPrivateKey(SERVER_0_KEY_FILE),
+        ImmutableList.of(getCertFromResourceName(SERVER_0_PEM_FILE)));
     assertThat(provider.savedKey).isNotNull();
     assertThat(provider.savedCertChain).isNotNull();
     assertThat(provider.getSslContext()).isNull();
@@ -145,14 +145,14 @@ public class CertProviderClientSslContextProviderTest {
     TestCallback testCallback =
         CommonTlsContextTestsUtil.getValueThruCallback(provider);
 
-    doChecksOnSslContext(false, testCallback.updatedSslContext, /* expectedApnProtos= */ null);
+    doChecksOnSslContext(true, testCallback.updatedSslContext, /* expectedApnProtos= */ null);
     TestCallback testCallback1 =
         CommonTlsContextTestsUtil.getValueThruCallback(provider);
     assertThat(testCallback1.updatedSslContext).isSameInstanceAs(testCallback.updatedSslContext);
 
     // just do root cert update: sslContext should still be the same
     watcherCaptor[0].updateTrustedRoots(
-        ImmutableList.of(getCertFromResourceName(SERVER_0_PEM_FILE)));
+        ImmutableList.of(getCertFromResourceName(CLIENT_PEM_FILE)));
     assertThat(provider.savedKey).isNull();
     assertThat(provider.savedCertChain).isNull();
     assertThat(provider.savedTrustedRoots).isNotNull();
@@ -172,18 +172,24 @@ public class CertProviderClientSslContextProviderTest {
   }
 
   @Test
-  public void testProviderForClient_mtls_newXds() throws Exception {
+  public void testProviderForServer_mtls_newXds() throws Exception {
     final CertificateProvider.DistributorWatcher[] watcherCaptor =
             new CertificateProvider.DistributorWatcher[1];
     TestCertificateProvider.createAndRegisterProviderProvider(
             certificateProviderRegistry, watcherCaptor, "testca", 0);
-    CertProviderClientSslContextProvider provider =
+    @SuppressWarnings("deprecation")
+    CertificateValidationContext staticCertValidationContext =
+        CertificateValidationContext.newBuilder().addAllMatchSubjectAltNames(Arrays
+            .asList(StringMatcher.newBuilder().setExact("foo.com").build(),
+                StringMatcher.newBuilder().setExact("bar.com").build())).build();
+    CertProviderServerSslContextProvider provider =
             getNewSslContextProvider(
                     "gcp_id",
                     "gcp_id",
                     CommonBootstrapperTestUtils.getTestBootstrapInfo(),
                     /* alpnProtocols= */ null,
-                    /* staticCertValidationContext= */ null);
+                    staticCertValidationContext,
+                    /* requireClientCert= */ true);
 
     assertThat(provider.savedKey).isNull();
     assertThat(provider.savedCertChain).isNull();
@@ -192,8 +198,8 @@ public class CertProviderClientSslContextProviderTest {
 
     // now generate cert update
     watcherCaptor[0].updateCertificate(
-            CommonCertProviderTestUtils.getPrivateKey(CLIENT_KEY_FILE),
-            ImmutableList.of(getCertFromResourceName(CLIENT_PEM_FILE)));
+            CommonCertProviderTestUtils.getPrivateKey(SERVER_0_KEY_FILE),
+            ImmutableList.of(getCertFromResourceName(SERVER_0_PEM_FILE)));
     assertThat(provider.savedKey).isNotNull();
     assertThat(provider.savedCertChain).isNotNull();
     assertThat(provider.getSslContext()).isNull();
@@ -208,14 +214,14 @@ public class CertProviderClientSslContextProviderTest {
     TestCallback testCallback =
             CommonTlsContextTestsUtil.getValueThruCallback(provider);
 
-    doChecksOnSslContext(false, testCallback.updatedSslContext, /* expectedApnProtos= */ null);
+    doChecksOnSslContext(true, testCallback.updatedSslContext, /* expectedApnProtos= */ null);
     TestCallback testCallback1 =
             CommonTlsContextTestsUtil.getValueThruCallback(provider);
     assertThat(testCallback1.updatedSslContext).isSameInstanceAs(testCallback.updatedSslContext);
 
     // just do root cert update: sslContext should still be the same
     watcherCaptor[0].updateTrustedRoots(
-            ImmutableList.of(getCertFromResourceName(SERVER_0_PEM_FILE)));
+            ImmutableList.of(getCertFromResourceName(CLIENT_PEM_FILE)));
     assertThat(provider.savedKey).isNull();
     assertThat(provider.savedCertChain).isNull();
     assertThat(provider.savedTrustedRoots).isNotNull();
@@ -235,18 +241,19 @@ public class CertProviderClientSslContextProviderTest {
   }
 
   @Test
-  public void testProviderForClient_queueExecutor() throws Exception {
+  public void testProviderForServer_queueExecutor() throws Exception {
     final CertificateProvider.DistributorWatcher[] watcherCaptor =
         new CertificateProvider.DistributorWatcher[1];
     TestCertificateProvider.createAndRegisterProviderProvider(
         certificateProviderRegistry, watcherCaptor, "testca", 0);
-    CertProviderClientSslContextProvider provider =
+    CertProviderServerSslContextProvider provider =
         getSslContextProvider(
             "gcp_id",
             "gcp_id",
             CommonBootstrapperTestUtils.getTestBootstrapInfo(),
             /* alpnProtocols= */ null,
-            /* staticCertValidationContext= */ null);
+            /* staticCertValidationContext= */ null,
+            /* requireClientCert= */ true);
     QueuedExecutor queuedExecutor = new QueuedExecutor();
 
     TestCallback testCallback =
@@ -255,8 +262,8 @@ public class CertProviderClientSslContextProviderTest {
 
     // now generate cert update
     watcherCaptor[0].updateCertificate(
-        CommonCertProviderTestUtils.getPrivateKey(CLIENT_KEY_FILE),
-        ImmutableList.of(getCertFromResourceName(CLIENT_PEM_FILE)));
+        CommonCertProviderTestUtils.getPrivateKey(SERVER_0_KEY_FILE),
+        ImmutableList.of(getCertFromResourceName(SERVER_0_PEM_FILE)));
     assertThat(queuedExecutor.runQueue).isEmpty(); // still empty
 
     // now generate root cert update
@@ -264,30 +271,34 @@ public class CertProviderClientSslContextProviderTest {
     assertThat(queuedExecutor.runQueue).hasSize(1);
     queuedExecutor.drain();
 
-    doChecksOnSslContext(false, testCallback.updatedSslContext, /* expectedApnProtos= */ null);
+    doChecksOnSslContext(true, testCallback.updatedSslContext, /* expectedApnProtos= */ null);
   }
 
   @Test
-  public void testProviderForClient_tls() throws Exception {
+  public void testProviderForServer_tls() throws Exception {
     final CertificateProvider.DistributorWatcher[] watcherCaptor =
         new CertificateProvider.DistributorWatcher[1];
     TestCertificateProvider.createAndRegisterProviderProvider(
         certificateProviderRegistry, watcherCaptor, "testca", 0);
-    CertProviderClientSslContextProvider provider =
+    CertProviderServerSslContextProvider provider =
         getSslContextProvider(
-            /* certInstanceName= */ null,
             "gcp_id",
+            /* rootInstanceName= */ null,
             CommonBootstrapperTestUtils.getTestBootstrapInfo(),
             /* alpnProtocols= */ null,
-            /* staticCertValidationContext= */ null);
+            /* staticCertValidationContext= */ null,
+            /* requireClientCert= */ false);
 
     assertThat(provider.savedKey).isNull();
     assertThat(provider.savedCertChain).isNull();
     assertThat(provider.savedTrustedRoots).isNull();
     assertThat(provider.getSslContext()).isNull();
 
-    // now generate root cert update
-    watcherCaptor[0].updateTrustedRoots(ImmutableList.of(getCertFromResourceName(CA_PEM_FILE)));
+    // now generate cert update
+    watcherCaptor[0].updateCertificate(
+            CommonCertProviderTestUtils.getPrivateKey(SERVER_0_KEY_FILE),
+            ImmutableList.of(getCertFromResourceName(SERVER_0_PEM_FILE)));
+
     assertThat(provider.getSslContext()).isNotNull();
     assertThat(provider.savedKey).isNull();
     assertThat(provider.savedCertChain).isNull();
@@ -296,11 +307,11 @@ public class CertProviderClientSslContextProviderTest {
     TestCallback testCallback =
         CommonTlsContextTestsUtil.getValueThruCallback(provider);
 
-    doChecksOnSslContext(false, testCallback.updatedSslContext, /* expectedApnProtos= */ null);
+    doChecksOnSslContext(true, testCallback.updatedSslContext, /* expectedApnProtos= */ null);
   }
 
   @Test
-  public void testProviderForClient_sslContextException_onError() throws Exception {
+  public void testProviderForServer_sslContextException_onError() throws Exception {
     CertificateValidationContext staticCertValidationContext =
             CertificateValidationContext.newBuilder()
                     .setTrustedCa(DataSource.newBuilder().setInlineString("foo"))
@@ -310,13 +321,19 @@ public class CertProviderClientSslContextProviderTest {
             new CertificateProvider.DistributorWatcher[1];
     TestCertificateProvider.createAndRegisterProviderProvider(
             certificateProviderRegistry, watcherCaptor, "testca", 0);
-    CertProviderClientSslContextProvider provider =
+    CertProviderServerSslContextProvider provider =
             getSslContextProvider(
-                    /* certInstanceName= */ null,
-                    "gcp_id",
+                    /* certInstanceName= */ "gcp_id",
+                    /* rootInstanceName= */ "gcp_id",
                     CommonBootstrapperTestUtils.getTestBootstrapInfo(),
                     /* alpnProtocols= */null,
-                    staticCertValidationContext);
+                    staticCertValidationContext,
+                    /* requireClientCert= */ true);
+
+    // now generate cert update
+    watcherCaptor[0].updateCertificate(
+            CommonCertProviderTestUtils.getPrivateKey(SERVER_0_KEY_FILE),
+            ImmutableList.of(getCertFromResourceName(SERVER_0_PEM_FILE)));
 
     TestCallback testCallback = new TestCallback(MoreExecutors.directExecutor());
     provider.addCallback(testCallback);
@@ -336,7 +353,7 @@ public class CertProviderClientSslContextProviderTest {
   }
 
   @Test
-  public void testProviderForClient_rootInstanceNull_expectError() throws Exception {
+  public void testProviderForServer_certInstanceNull_expectError() throws Exception {
     final CertificateProvider.DistributorWatcher[] watcherCaptor =
         new CertificateProvider.DistributorWatcher[1];
     TestCertificateProvider.createAndRegisterProviderProvider(
@@ -347,33 +364,11 @@ public class CertProviderClientSslContextProviderTest {
           /* rootInstanceName= */ null,
           CommonBootstrapperTestUtils.getTestBootstrapInfo(),
           /* alpnProtocols= */ null,
-          /* staticCertValidationContext= */ null);
+          /* staticCertValidationContext= */ null,
+          /* requireClientCert= */ false);
       fail("exception expected");
     } catch (NullPointerException expected) {
-      assertThat(expected).hasMessageThat().contains("Client SSL requires rootCertInstance");
+      assertThat(expected).hasMessageThat().contains("Server SSL requires certInstance");
     }
   }
-
-  static class QueuedExecutor implements Executor {
-    /** A list of Runnables to be run in order. */
-    @VisibleForTesting final Queue<Runnable> runQueue = new ConcurrentLinkedQueue<>();
-
-    @Override
-    public synchronized void execute(Runnable r) {
-      runQueue.add(checkNotNull(r, "'r' must not be null."));
-    }
-
-    public synchronized void drain() {
-      Runnable r;
-      while ((r = runQueue.poll()) != null) {
-        try {
-          r.run();
-        } catch (RuntimeException e) {
-          // Log it and keep going.
-          logger.log(Level.SEVERE, "Exception while executing runnable " + r, e);
-        }
-      }
-    }
-  }
-
 }
